sell/src/main/java/com/chen/service/impl/SellServiceImpl.java
package com.chen.service.impl;

import com.chen.dataobject.SellerInfo;
import com.chen.repository.SellerInfoRepository;
import com.chen.service.SellerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class SellServiceImpl implements SellerService {

    @Autowired
    private SellerInfoRepository repository;

    @Override
    public SellerInfo findSellerInfoByOpenid(String openid) {
        return repository.findByOpenid(openid);
    }
}
