sell/src/test/java/com/chen/service/impl/SellServiceImplTest.java
package com.chen.service.impl;

import com.chen.dataobject.SellerInfo;
import lombok.extern.slf4j.Slf4j;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.*;
@RunWith(SpringRunner.class)
@SpringBootTest
@Slf4j
public class SellServiceImplTest {

    private static final String openid = "abc";

    @Autowired
    private SellServiceImpl sellService;

    @Test
    public void findSellerInfoByOpenid() {
        SellerInfo result = sellService.findSellerInfoByOpenid(openid);
        Assert.assertEquals(openid,result.getOpenid());
    }
}